third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/prelu.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] prelu float32 5D tensors
  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'prelu' on 'MLGraphBuilder': Unsupported rank 5 for argument input (must be at most 4)."
[FAIL] prelu float32 broadcast 4D x 1D slope of shape [1]
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': The input and slope should have the same last dimension."
[FAIL] prelu float32 broadcast 4D x 4D slope
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': The input and slope should have the same last dimension."
[FAIL] prelu float16 0D scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual 0 should be close enough to expected -5.3671875 by ULP distance: expected a number less than or equal to 1 but got 17758
[FAIL] prelu float16 1D constant tensors
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual -15232 should be close enough to expected -23.8125 by ULP distance: expected a number less than or equal to 1 but got 9596
[FAIL] prelu float16 1D tensors
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual -15232 should be close enough to expected -23.8125 by ULP distance: expected a number less than or equal to 1 but got 9596
[FAIL] prelu float16 1D non-constant slope
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual -0.0003886222839355469 should be close enough to expected -23.8125 by ULP distance: expected a number less than or equal to 1 but got 16278
[FAIL] prelu float16 2D tensors
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual -15232 should be close enough to expected -23.8125 by ULP distance: expected a number less than or equal to 1 but got 9596
[FAIL] prelu float16 3D tensors
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual -15232 should be close enough to expected -23.8125 by ULP distance: expected a number less than or equal to 1 but got 9596
[FAIL] prelu float16 4D tensors
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual -15232 should be close enough to expected -23.8125 by ULP distance: expected a number less than or equal to 1 but got 9596
[FAIL] prelu float16 5D tensors
  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'prelu' on 'MLGraphBuilder': Unsupported rank 5 for argument input (must be at most 4)."
[FAIL] prelu float16 broadcast 4D x 1D slope
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual 2 should be close enough to expected -12.9296875 by ULP distance: expected a number less than or equal to 1 but got 35447
[FAIL] prelu float16 broadcast 4D x 1D slope of shape [1]
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': The input and slope should have the same last dimension."
[FAIL] prelu float16 broadcast 4D x 2D slope
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual -7948 should be close enough to expected -12.421875 by ULP distance: expected a number less than or equal to 1 but got 9613
[FAIL] prelu float16 broadcast 4D x 3D slope
  assert_less_than_equal: assert_array_approx_equals_ulp: test prelu float16 actual 2 should be close enough to expected -12.9296875 by ULP distance: expected a number less than or equal to 1 but got 35447
[FAIL] prelu float16 broadcast 4D x 4D slope
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': The input and slope should have the same last dimension."
[FAIL] prelu int64 2D constant tensors
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type int64 must be one of [float32,float16]."
Harness: the test ran to completion.
